OLD_CODE/experimental/new_dataset/requirements.txt
scikit_image==0.16.1
numpy==1.15.4
easydict==1.9
torchvision==0.2.1
pandas==0.24.2
matplotlib==3.1.1
tabulate==0.8.3
torch==1.2.0
python_gflags==3.1.2
skimage==0.0
PyYAML==5.1.2
